backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Config
pydantic-settings==2.2.1
python-dotenv==1.0.1

# PDF parsing
pdfplumber==0.11.0
pdf2image==1.17.0

# OCR fallback
pytesseract==0.3.10
Pillow==10.3.0

# LLM - Groq
groq==0.9.0

# Embeddings
sentence-transformers==3.0.1

# Vector store
chromadb==0.5.0

# Database
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
